+++ src/main/java/ktb3/full/week04/repository/base/PagingAndSortingRepository.java
@@ -1,4 +1,4 @@
-package ktb3.full.week04.repository;
+package ktb3.full.week04.repository.base;
 
 import ktb3.full.week04.dto.page.PageRequest;
 import ktb3.full.week04.dto.page.PageResponse;
